requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.9.1
aiosqlite==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
